+++ packages/cli/__tests__/zosfiles/__system__/create/pds/cli.files.create.dataset.pds.system.test.ts
@@ -10,11 +10,12 @@
 */
 
 import { Session } from "@zowe/imperative";
-import { ITestEnvironment, runCliScript } from "@zowe/cli-test-utils";
 import { TestEnvironment } from "../../../../../../../__tests__/__src__/environment/TestEnvironment";
+import { ITestEnvironment } from "../../../../../../../__tests__/__src__/environment/ITestEnvironment";
 import { ITestPropertiesSchema } from "../../../../../../../__tests__/__src__/properties/ITestPropertiesSchema";
-import { stripNewLines, delay, delTime } from "../../../../../../../__tests__/__src__/TestUtils";
+import { stripNewLines, wait, waitTime } from "../../../../../../../__tests__/__src__/TestUtils";
 import { Delete } from "@zowe/zos-files-for-zowe-sdk";
+import { runCliScript } from "@zowe/cli-test-utils";
 
 const ZOWE_OPT_BASE_PATH = "ZOWE_OPT_BASE_PATH";
 
@@ -68,7 +69,7 @@ describe("Create Partitioned Data Set", () => {
         afterEach(async () => {
             // use DELETE APIs
             if (dsnameSuffix !== "") {
-                await delay(delTime);
+                await wait(waitTime); //wait 2 seconds
                 const response = await Delete.dataSet(REAL_SESSION, dsname + "." + dsnameSuffix);
             }
         });
@@ -104,7 +105,7 @@ describe("Create Partitioned Data Set", () => {
         afterEach(async () => {
             // use DELETE APIs
             if (dsnameSuffix !== "") {
-                await delay(delTime);
+                await wait(waitTime); //wait 2 seconds
                 const response = await Delete.dataSet(REAL_SESSION, dsname + "." + dsnameSuffix);
             }
         });