+++ packages/cli/__tests__/zosjobs/__system__/download/__scripts__/download-output/download_fully_qualified.sh
@@ -22,6 +22,9 @@ then
     exit $RC
 fi
 
+# Echo the job ID for further use
+echo "Submitted job ID: $JOBID"
+
 # Loop until the job goes to the output queue
 until [ $ATTEMPTS -gt 0 ]
 do