+++ packages/cli/__tests__/zosjobs/__system__/search/cli.zos-jobs.search.job.system.test.ts
@@ -9,23 +9,23 @@
 *
 */
 
-import { ITestEnvironment, runCliScript } from "@zowe/cli-test-utils";
+import { runCliScript } from "@zowe/cli-test-utils";
 import { TestEnvironment } from "../../../../../../__tests__/__src__/environment/TestEnvironment";
 import { ITestPropertiesSchema } from "../../../../../../__tests__/__src__/properties/ITestPropertiesSchema";
 import { Session } from "@zowe/imperative";
+import { ITestEnvironment } from "../../../../../../__tests__/__src__/environment/ITestEnvironment";
+import { SubmitJobs } from "@zowe/zos-jobs-for-zowe-sdk";
 
 // Test Environment populated in the beforeAll();
 let TEST_ENVIRONMENT: ITestEnvironment<ITestPropertiesSchema>;
-let IEFBR14_JOB: string;
 let REAL_SESSION: Session;
-let ACCOUNT: string;
 let JOB_NAME: string;
-let NON_HELD_JOBCLASS: string;
-let SEARCH_STRING: string;
-let REGEX_STRING: string;
-let BAD_SEARCH_STRING: string;
+const SEARCH_STRING = "PGM=IEFBR14";
+const REGEX_STRING = "IEFBR14|RC=0000";
+const BAD_SEARCH_STRING = "bluhbluh";
 
 describe("zos-jobs search job command", () => {
+
     // Create the unique test environment
     beforeAll(async () => {
         TEST_ENVIRONMENT = await TestEnvironment.setUp({
@@ -33,18 +33,12 @@ describe("zos-jobs search job command", () => {
             tempProfileTypes: ["zosmf"]
         });
 
-        IEFBR14_JOB = TEST_ENVIRONMENT.systemTestProperties.zosjobs.iefbr14Member;
-        const defaultSystem = TEST_ENVIRONMENT.systemTestProperties;
         REAL_SESSION = TestEnvironment.createZosmfSession(TEST_ENVIRONMENT);
+        const iefbr14Job = TEST_ENVIRONMENT.systemTestProperties.zosjobs.iefbr14Member;
 
-        ACCOUNT = defaultSystem.tso.account;
-        const JOB_LENGTH = 6;
-        JOB_NAME = REAL_SESSION.ISession.user.substring(0, JOB_LENGTH).toUpperCase() + "SF";
-        NON_HELD_JOBCLASS = TEST_ENVIRONMENT.systemTestProperties.zosjobs.jobclass;
-        SEARCH_STRING = "PGM=IEFBR14";
-        REGEX_STRING = "IEFBR14|RC=0000";
-        BAD_SEARCH_STRING = "bluhbluh";
-
+        const job = await SubmitJobs.submitJob(REAL_SESSION, iefbr14Job);
+        JOB_NAME = job.jobname;
+        TEST_ENVIRONMENT.resources.jobs.push(job);
     });
 
     afterAll(async () => {
@@ -53,17 +47,16 @@ describe("zos-jobs search job command", () => {
 
     describe("Successful response", () => {
         it("should be able to search for a string in every spool file for a job", () => {
-            const response = runCliScript(__dirname + "/__scripts__/job/search_string_spool_content.sh",
-                TEST_ENVIRONMENT, [IEFBR14_JOB, JOB_NAME, SEARCH_STRING]);
+            const response = runCliScript(__dirname + "/__scripts__/job/search_job_string.sh",
+                TEST_ENVIRONMENT, [JOB_NAME, SEARCH_STRING]);
             expect(response.stderr.toString()).toBe("");
             expect(response.status).toBe(0);
             expect(response.stdout.toString()).toContain(SEARCH_STRING);
         });
 
         it("should be able to search for a regex in every spool file for a job", () => {
-            const argString = "--search-regex \"" + REGEX_STRING + "\"";
-            const response = runCliScript(__dirname + "/__scripts__/job/search_no_job_submit.sh",
-                TEST_ENVIRONMENT, [JOB_NAME, argString]);
+            const response = runCliScript(__dirname + "/__scripts__/job/search_job_regex.sh",
+                TEST_ENVIRONMENT, [JOB_NAME, REGEX_STRING]);
             expect(response.stderr.toString()).toBe("");
             expect(response.status).toBe(0);
             expect(response.stdout.toString()).toContain("IEFBR14");
@@ -71,9 +64,8 @@ describe("zos-jobs search job command", () => {
         });
 
         it("should limit the search when the --search-limit and --file-limit options are specified", () => {
-            const argString = "--search-string \"" + SEARCH_STRING + "\" --search-limit 5 --file-limit 3";
-            const response = runCliScript(__dirname + "/__scripts__/job/search_no_job_submit.sh",
-                TEST_ENVIRONMENT, [JOB_NAME, argString]);
+            const response = runCliScript(__dirname + "/__scripts__/job/search_job_string.sh",
+                TEST_ENVIRONMENT, [JOB_NAME, SEARCH_STRING, "--search-limit", "5", "--file-limit", "3"]);
             expect(response.stderr.toString()).toBe("");
             expect(response.status).toBe(0);
             expect(response.stdout.toString()).toContain(SEARCH_STRING);
@@ -82,25 +74,23 @@ describe("zos-jobs search job command", () => {
 
     describe("error handling", () => {
         it("should return a status code of 1 if the string is not found", () => {
-            const argString = "--search-string \"" + BAD_SEARCH_STRING + "\"";
-            const response = runCliScript(__dirname + "/__scripts__/job/search_no_job_submit.sh",
-                TEST_ENVIRONMENT, [JOB_NAME, argString]);
+            const response = runCliScript(__dirname + "/__scripts__/job/search_job_string.sh",
+                TEST_ENVIRONMENT, [JOB_NAME, BAD_SEARCH_STRING]);
             expect(response.stderr.toString()).toContain("The search spool job command returned a non-zero rc: 1");
             expect(response.status).toBe(1);
             expect(response.stdout.toString()).toBe("");
         });
 
         it("should fail if no parameters are passed", () => {
-            const response = runCliScript(__dirname + "/__scripts__/job/search_no_job_submit.sh",
-                TEST_ENVIRONMENT, [JOB_NAME, ""]);
+            const response = runCliScript(__dirname + "/__scripts__/job/missing_search_string.sh",
+                TEST_ENVIRONMENT, [JOB_NAME]);
             expect(response.stderr.toString()).toContain("You must specify either the `--search-string` or `--search-regex` option");
             expect(response.status).toBe(1);
         });
 
         it("should fail if --search-string and --search-regex parameters are passed", () => {
-            const argString = "--search-regex \"" + REGEX_STRING +"\" --search-string \"" + SEARCH_STRING + "\"";
-            const response = runCliScript(__dirname + "/__scripts__/job/search_no_job_submit.sh",
-                TEST_ENVIRONMENT, [JOB_NAME, argString]);
+            const response = runCliScript(__dirname + "/__scripts__/job/search_job_string.sh",
+                TEST_ENVIRONMENT, [JOB_NAME, SEARCH_STRING, "--search-regex", REGEX_STRING]);
             expect(response.stderr.toString()).toContain("You must specify either the `--search-string` or `--search-regex` option");
             expect(response.status).toBe(1);
         });