+++ packages/cli/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
-
 All notable changes to the Zowe CLI package will be documented in this file.
+
 ## Recent Changes
 - Enhancement: Add --ignore-not-found flag logic to zosfilesBase.handler and update command definitions and en.ts for delete operations. [#2254](https://github.com/zowe/zowe-cli/pull/2254)